Test passed.

+++ app/radius/radmainwindow.h
@@ -46,11 +46,11 @@ private:
     QAction * actCalc1;
     QAction * actCalc2;
 
-    complex<long double> * stc;// = new complex<long double> [nd];
-    complex<long double> * stc1;/// = new complex<long double> [nd];
-    long double * stc2;
-    complex<long double> * stc3;
-    complex<long double> * stc4;// = new complex<double> [nd];
+    complex<double> * stc;// = new complex<double> [nd];
+    complex<double> * stc1;/// = new complex<double> [nd];
+    double * stc2;
+    complex<double> * stc3;
+    complex<double> * stc4;// = new complex<double> [nd];
 private:
     Q_OBJECT
 };
