+++ src/calcim.h
@@ -13,7 +13,7 @@ public:
     Calc ();
     virtual ~Calc ();
 
-    virtual long double* operator () (const double* st, int nd) = 0;
+    virtual double* operator () (const double* st, int nd) = 0;
 };
 
 class _FFTMATR_EXPORT_ Calc1 : public Calc
@@ -22,7 +22,7 @@ public:
     Calc1 ();
     virtual ~Calc1 ();
 
-    virtual long double * operator () (const double* st, int nd);
+    virtual double * operator () (const double* st, int nd);
 };
 
 class _FFTMATR_EXPORT_ Calc2 : public Calc
@@ -31,7 +31,7 @@ public:
     Calc2 ();
     virtual ~Calc2 ();
 
-    virtual long double * operator () (const double* st, int nd);
+    virtual double * operator () (const double* st, int nd);
 };
 
 #endif
