+++ src/fft_c.h
@@ -14,7 +14,7 @@ class _FFTMATR_EXPORT_ FFT_Transform
 public:
     FFT_Transform (void);
     virtual ~FFT_Transform (void);
-    virtual complex<long double> * operator () (complex<long double> * src, int nsrc, int n2, int sign, unsigned flags);
+    virtual complex<double> * operator () (complex<double> * src, int nsrc, int n2, int sign, unsigned flags);
 
     static int pow2roundup (int n);
 };
@@ -24,7 +24,7 @@ class _FFTMATR_EXPORT_ FFT_RealTransform
 public:
     FFT_RealTransform (void);
     virtual ~FFT_RealTransform (void);
-    virtual complex<long double> * operator () (long double * src, int nsrc, int n2, int sign, unsigned flags);
+    virtual complex<double> * operator () (double * src, int nsrc, int n2, int sign, unsigned flags);
 };
 
 class _FFTMATR_EXPORT_ FFT2_Transform
@@ -32,7 +32,7 @@ class _FFTMATR_EXPORT_ FFT2_Transform
 public:
     FFT2_Transform (void);
     virtual ~FFT2_Transform (void);
-    virtual complex<long double> * operator () (complex<long double> * src, int nr, int nc, int sign, unsigned flags);
+    virtual complex<double> * operator () (complex<double> * src, int nr, int nc, int sign, unsigned flags);
 
 };
