+++ src/matrix.h
@@ -14,16 +14,16 @@ class _FFTMATR_EXPORT_ CMatrix
 public:
     CMatrix (int nrows=0, int ncols=0);
     CMatrix (double value, int nrows, int ncols);
-    CMatrix (complex<long double> cval, int nrows, int ncols);
+    CMatrix (complex<double> cval, int nrows, int ncols);
     ~CMatrix (void);
 
-    complex<long double>& operator() (int i, int j);
-    const complex<long double>& operator() (int i, int j) const;
-    complex<long double> * getData (void) const;
+    complex<double>& operator() (int i, int j);
+    const complex<double>& operator() (int i, int j) const;
+    complex<double> * getData (void) const;
 private:
     int nRows;
     int nCols;
-    complex<long double> * matr;
+    complex<double> * matr;
 };
 
 #endif
